TFP-4747: Lar være å legge inn ekstra linjeskift midt i setninger, da copy-paste fra PDF gir linjeskift ved line wrap. (#1574)
TFP-4751: Sørger for at avrunding av beløp blir riktig i innvilgelse SVP.

+++ domenetjenester/brevbestiller/src/main/java/no/nav/foreldrepenger/melding/brevmapper/brev/innvilgelsesvp/SvangerskapspengerInnvilgelseDokumentdataMapper.java
@@ -117,7 +117,7 @@ public class SvangerskapspengerInnvilgelseDokumentdataMapper implements Dokument
             dokumentdataBuilder.medSelvstendigNæringsdrivende(mapSelvstendigNæringsdrivende(beregningsgrunnlag));
             dokumentdataBuilder.medFrilanser(mapFrilanser(beregningsgrunnlag));
             dokumentdataBuilder.medNaturalytelser(mapNaturalytelser(beregningsresultat, beregningsgrunnlag, språkkode));
-            dokumentdataBuilder.medBruttoBeregningsgrunnlag(getAvkortetPrÅrSVP(beregningsgrunnlag).longValue());
+            dokumentdataBuilder.medBruttoBeregningsgrunnlag(getAvkortetPrÅrSVP(beregningsgrunnlag));
             dokumentdataBuilder.medMilitærSivil(erMilitærSivil(beregningsgrunnlag));
             dokumentdataBuilder.medInntektOver6G(inntektOverSeksG(beregningsgrunnlag));
             dokumentdataBuilder.medSeksG(finnSeksG(beregningsgrunnlag).longValue());
